requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
redis==5.0.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
